server/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
chardet==5.2.0
click==8.2.1
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.10.0
moviepy==2.2.1
numpy==2.3.2
openai==1.101.0
pillow==11.3.0
proglog==0.1.12
pydantic==2.11.7
pydantic_core==2.33.2
pysrt==1.1.2
python-dotenv==1.1.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.47.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
